fix: Use CSS grid for home page header
The handrolled col-* classes were difficult to reason about and caused
problems on the home page header.

Replace them with a CSS grid that splits the header into thirds, and
gives over the first two thirds of each row to a primary content column,
and the remaining third to an additional content column.

Put the name and device screenshot in the first column, the logo and
links in the second column.